# openapi.yaml
openapi: 3.1.0
info:
  title: Dataminer
  description: Document extraction service for legal documents
  version: 0.1.0
paths:
...
  /v1/dataminer/fields/{field_id}:
    get:
      tags:
      - fields
      summary: Get field details
      description: Get detailed information about a specific field definition
      operationId: get_field_v1_dataminer_fields__field_id__get
      parameters:
      - name: field_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Field Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldDefinitionResponse'
        '404':
          description: Field not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - fields
      summary: Update field definition
      description: Update an existing field definition
      operationId: update_field_v1_dataminer_fields__field_id__put
      parameters:
      - name: field_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Field Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FieldDefinitionUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldDefinitionResponse'
        '404':
          description: Field not found
        '409':
          description: Field with same name already exists
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - fields
      summary: Delete field definition
      description: Delete a field definition
      operationId: delete_field_v1_dataminer_fields__field_id__delete
      parameters:
      - name: field_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Field Id
      responses:
        '204':
          description: Field deleted successfully
        '404':
          description: Field not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    FieldDefinitionResponse:
      properties:
        field_id:
          type: string
          format: uuid
          title: Field Id
        source_id:
          type: string
          title: Source Id
        field_name:
          type: string
          maxLength: 100
          title: Field Name
          description: Internal field name (snake_case)
        field_display_name:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Field Display Name
          description: Human-readable display name
        field_category:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Field Category
          description: Category grouping (metadata, parties, decision, dates)
        field_type:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Field Type
          description: Data type (string, date, number, array, object)
        extraction_method:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Extraction Method
          description: How to extract this field (regex, llm, hybrid, rule_based)
        extraction_section:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Extraction Section
          description: Document section to search for this field
        regex_pattern:
          anyOf:
          - type: string
          - type: 'null'
          title: Regex Pattern
          description: Regular expression pattern for regex extraction
        llm_prompt_template_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Llm Prompt Template Id
          description: Reference to prompt template for LLM extraction
        is_required:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Required
          description: Whether this field is required for valid extraction
          default: false
        validation_rules:
          anyOf:
          - type: object
          - type: 'null'
          title: Validation Rules
          description: JSON object with validation rules for the field
        confidence_threshold:
          anyOf:
          - type: number
            minimum: 0.0
            maximum: 1.0
          - type: string
            pattern: ^[+-]?\d+(\.\d{1,2})?$
          - type: 'null'
          title: Confidence Threshold
          description: Minimum confidence score to accept extraction (0.00-1.00)
          default: '0.75'
        normalization_rules:
          anyOf:
          - type: object
          - type: 'null'
          title: Normalization Rules
          description: JSON object with normalization/transformation rules
        display_order:
          anyOf:
          - type: integer
          - type: 'null'
          title: Display Order
          description: Order in which to display this field in UI
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - field_id
      - source_id
      - field_name
      title: FieldDefinitionResponse
      description: Schema for field definition response.
...
    FieldDefinitionUpdate:
      properties:
        field_name:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Field Name
          description: Internal field name (snake_case)
        field_display_name:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Field Display Name
          description: Human-readable display name
        field_category:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Field Category
          description: Category grouping (metadata, parties, decision, dates)
        field_type:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Field Type
          description: Data type (string, date, number, array, object)
        extraction_method:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Extraction Method
          description: How to extract this field (regex, llm, hybrid, rule_based)
        extraction_section:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Extraction Section
          description: Document section to search for this field
        regex_pattern:
          anyOf:
          - type: string
          - type: 'null'
          title: Regex Pattern
          description: Regular expression pattern for regex extraction
        llm_prompt_template_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Llm Prompt Template Id
          description: Reference to prompt template for LLM extraction
        is_required:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Required
          description: Whether this field is required for valid extraction
        validation_rules:
          anyOf:
          - type: object
          - type: 'null'
          title: Validation Rules
          description: JSON object with validation rules for the field
        confidence_threshold:
          anyOf:
          - type: number
            minimum: 0.0
            maximum: 1.0
          - type: string
            pattern: ^[+-]?\d+(\.\d{1,2})?$
          - type: 'null'
          title: Confidence Threshold
          description: Minimum confidence score to accept extraction (0.00-1.00)
        normalization_rules:
          anyOf:
          - type: object
          - type: 'null'
          title: Normalization Rules
          description: JSON object with normalization/transformation rules
        display_order:
          anyOf:
          - type: integer
          - type: 'null'
          title: Display Order
          description: Order in which to display this field in UI
      type: object
      title: FieldDefinitionUpdate
      description: Schema for updating a field definition.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError